adding ability for spackbot to suggest opening multiple PRs for more than one package
