+++ spackbot/handlers/__init__.py
@@ -2,3 +2,4 @@ from .pipelines import run_pipeline
 from .labels import add_labels
 from .reviewers import add_reviewers
 from .style import style_comment
+from .packages import count_packages
